requirements.txt
asgiref==3.7.2
Django==3.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
PyJWT==2.8.0
pytz==2023.3.post1
sqlparse==0.4.4
typing-extensions==4.7.1
gunicorn==20.1.0
django-environ==0.4.5
psycopg2-binary==2.9.2
django-cors-headers==4.1.0